public_api/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
